pyproject.toml
[project]
name = "acs-10x-chemist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "imageio>=2.37.0",
    "modal>=1.1.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rdkit>=2025.3.5",
]
